requirements.txt
connexion
python-jose
flake8